# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhaleWatch API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/investors": {
      "get": {
        "tags": [
          "investors"
        ],
        "summary": "Every tracked investor, with its latest published portfolio",
        "description": "List investors, paginated. Each row names its own ``latest_period``.",
        "operationId": "listInvestors",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/InvestorSort",
              "description": "`value` and `positions` largest first, as of each investor's latest period. `name` A to Z. Investors with nothing published come last under the first two.",
              "default": "value"
            },
            "description": "`value` and `positions` largest first, as of each investor's latest period. `name` A to Z. Investors with nothing published come last under the first two."
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FilerCategory"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only investors of this style.",
              "title": "Category"
            },
            "description": "Only investors of this style."
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "Part of the investor's name, its name on EDGAR, or its manager's.",
              "examples": [
                "buffett"
              ],
              "title": "Q"
            },
            "description": "Part of the investor's name, its name on EDGAR, or its manager's."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Rows per page. Above 200 is refused, not cut down.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Rows per page. Above 200 is refused, not cut down."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`page.next_cursor` from the previous page. Omit for the first page. Keep the other parameters the same while paging.",
              "title": "Cursor"
            },
            "description": "`page.next_cursor` from the previous page. Omit for the first page. Keep the other parameters the same while paging."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestorSummaryEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "`cursor` is malformed, or came from a different listing or sort order.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "A parameter is missing or malformed. `errors` says which.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/investors/{slug}": {
      "get": {
        "tags": [
          "investors"
        ],
        "summary": "One investor",
        "description": "One investor, with the list's fields and its concentration, turnover,\nfirst period and CIKs.",
        "operationId": "getInvestor",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The investor's slug, as `slug` on its row of `GET /v1/investors`.",
              "examples": [
                "berkshire-hathaway"
              ],
              "title": "Slug"
            },
            "description": "The investor's slug, as `slug` on its row of `GET /v1/investors`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestorDetail"
                }
              }
            }
          },
          "404": {
            "description": "No investor with that slug.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "A parameter is missing or malformed. `errors` says which.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "invalid_cursor",
          "not_found",
          "method_not_allowed",
          "validation_error",
          "http_error"
        ],
        "title": "ErrorCode",
        "description": "What went wrong, for a program to branch on. ``detail`` says it for a person."
      },
...
      "FilerCategory": {
        "type": "string",
        "enum": [
          "value",
          "activist",
          "quant",
          "macro",
          "growth",
          "multi_strategy"
        ],
        "title": "FilerCategory",
        "description": "The investment style a filer is filed under on the site.\n\nText with a ``CHECK``, not a native enum, by the rule in\n:mod:`app.db.models.enums`: the set is ours, it is editorial, and the day a\nseventh style is wanted should be one line of DDL rather than an\n``ALTER TYPE`` that cannot be reversed.\n\nOne per filer, and deliberately coarse. It exists to group pages and to cut\nthe aggregate views, not to describe a strategy — a style label precise\nenough to be uncontroversial would be too fine to group anything by."
      },
...
      "InvestorDetail": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "The investor's stable identifier: `/v1/investors/{slug}`.",
            "examples": [
              "berkshire-hathaway"
            ]
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "Our name for the institution, or the name on its latest cover page.",
            "examples": [
              "Berkshire Hathaway"
            ]
          },
          "manager_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manager Name",
            "description": "The person the institution is known by, where there is one.",
            "examples": [
              "Warren Buffett"
            ]
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FilerCategory"
              },
              {
                "type": "null"
              }
            ],
            "description": "The investment style we file it under. Null when we have not chosen one.",
            "examples": [
              "value"
            ]
          },
          "latest_period": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Period",
            "description": "The newest quarter end we have published for this filer, which every figure below describes. Null when nothing is published yet.",
            "examples": [
              "2026-06-30"
            ]
          },
          "last_filed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Filed At",
            "description": "When the most recently filed of the filings behind `latest_period` reached EDGAR: a later amendment, if one counts.",
            "examples": [
              "2026-08-14T16:32:05Z"
            ]
          },
          "portfolio_value_usd": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Value Usd",
            "description": "Whole dollars, common stock only: no options, no principal amounts.",
            "examples": [
              "263479238410"
            ]
          },
          "position_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Count",
            "description": "Common-stock positions in `latest_period`.",
            "examples": [
              41
            ]
          },
          "top_holding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TopHolding"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when `latest_period` is, and when the portfolio is empty."
          },
          "sparkline": {
            "items": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "decimal"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Sparkline",
            "description": "Portfolio value for the 8 quarters ending at `latest_period`, oldest first, so the last entry is `portfolio_value_usd`. Null for a quarter with nothing published. Empty when `latest_period` is null.",
            "examples": [
              [
                "231000000000",
                null,
                "240000000000",
                "263479238410"
              ]
            ]
          },
          "first_period": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Period",
            "description": "The oldest quarter end we have published for this filer.",
            "examples": [
              "2021-09-30"
            ]
          },
          "top10_weight_pct": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top10 Weight Pct",
            "description": "How much of `latest_period`'s value is in its ten largest positions, in %.",
            "examples": [
              "88.112045"
            ]
          },
          "turnover_pct": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnover Pct",
            "description": "Shares bought and sold in `latest_period`, valued at its prices, halved, as a percentage of the previous period's value. Null in the filer's first period, which has no previous one.",
            "examples": [
              "4.381002"
            ]
          },
          "ciks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ciks",
            "description": "Every CIK the filer files under, oldest entity first.",
            "examples": [
              [
                "0001067983"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "slug",
          "display_name",
          "sparkline",
          "ciks"
        ],
        "title": "InvestorDetail",
        "description": "One filer, with what the list leaves out."
      },
...
      "InvestorSort": {
        "type": "string",
        "enum": [
          "value",
          "positions",
          "name"
        ],
        "title": "InvestorSort",
        "description": "What the investor list is ordered by."
      },
      "InvestorSummary": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "The investor's stable identifier: `/v1/investors/{slug}`.",
            "examples": [
              "berkshire-hathaway"
            ]
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "Our name for the institution, or the name on its latest cover page.",
            "examples": [
              "Berkshire Hathaway"
            ]
          },
          "manager_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manager Name",
            "description": "The person the institution is known by, where there is one.",
            "examples": [
              "Warren Buffett"
            ]
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FilerCategory"
              },
              {
                "type": "null"
              }
            ],
            "description": "The investment style we file it under. Null when we have not chosen one.",
            "examples": [
              "value"
            ]
          },
          "latest_period": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Period",
            "description": "The newest quarter end we have published for this filer, which every figure below describes. Null when nothing is published yet.",
            "examples": [
              "2026-06-30"
            ]
          },
          "last_filed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Filed At",
            "description": "When the most recently filed of the filings behind `latest_period` reached EDGAR: a later amendment, if one counts.",
            "examples": [
              "2026-08-14T16:32:05Z"
            ]
          },
          "portfolio_value_usd": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Value Usd",
            "description": "Whole dollars, common stock only: no options, no principal amounts.",
            "examples": [
              "263479238410"
            ]
          },
          "position_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Count",
            "description": "Common-stock positions in `latest_period`.",
            "examples": [
              41
            ]
          },
          "top_holding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TopHolding"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when `latest_period` is, and when the portfolio is empty."
          },
          "sparkline": {
            "items": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "decimal"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Sparkline",
            "description": "Portfolio value for the 8 quarters ending at `latest_period`, oldest first, so the last entry is `portfolio_value_usd`. Null for a quarter with nothing published. Empty when `latest_period` is null.",
            "examples": [
              [
                "231000000000",
                null,
                "240000000000",
                "263479238410"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "slug",
          "display_name",
          "sparkline"
        ],
        "title": "InvestorSummary",
        "description": "One filer, as a row of the list."
      },
      "InvestorSummaryEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/InvestorSummary"
            },
            "type": "array",
            "title": "Data",
            "description": "The rows, in the order the endpoint documents."
          },
          "meta": {
            "$ref": "#/components/schemas/Meta",
            "description": "What the rows describe, and when this answer was built."
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page"
              },
              {
                "type": "null"
              }
            ],
            "description": "How to get the next page. Null when the endpoint is not paged."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "InvestorSummaryEnvelope",
        "description": "Every tracked investor, a row each."
      },
...
      "Page": {
        "properties": {
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "The page size this response was built with.",
            "examples": [
              50
            ]
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Pass as `?cursor=` for the next page, with the same filters and sort. Null on the last page. Opaque: its contents are not a contract.",
            "examples": [
              "eyJ2IjoxLCJrIjpbIjE3NDM0NjA0ODQ3Mi4wMCIsNDJdfQ"
            ]
          }
        },
        "type": "object",
        "required": [
          "limit"
        ],
        "title": "Page",
        "description": "Where this page ends, as an opaque token to send back for the next one."
      },
...
      "Problem": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What went wrong, stable across releases. `invalid_cursor`: `cursor` is malformed, or belongs to a different listing or sort. `not_found`: no such thing, or nothing published for the period asked for. `validation_error`: a parameter is missing or malformed; `errors` says which.",
            "examples": [
              "not_found"
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "What went wrong, in a sentence for a person. Not a contract: do not parse it.",
            "examples": [
              "No investor 'berkshire'. GET /v1/investors lists every one."
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "title": "Problem",
        "description": "Why the request was refused."
      },
...
      "ValidationProblem": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What went wrong, stable across releases. `invalid_cursor`: `cursor` is malformed, or belongs to a different listing or sort. `not_found`: no such thing, or nothing published for the period asked for. `validation_error`: a parameter is missing or malformed; `errors` says which.",
            "examples": [
              "not_found"
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "What went wrong, in a sentence for a person. Not a contract: do not parse it.",
            "examples": [
              "No investor 'berkshire'. GET /v1/investors lists every one."
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors",
            "description": "Every parameter that failed, each once. Never empty."
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail",
          "errors"
        ],
        "title": "ValidationProblem",
        "description": "The 422 body: a parameter is missing or malformed."
      }
    }
  }
}